Handle talk page title when empty page or no DISPLAYTITLE
Talk page title is "Talk:Page Name" instead of "Discuss Page Name/Display Title",
when talk page is empty (not created) or uses Flow/StructuredDiscussions
or when DISPLAYTITLE is not set.
We add a condition handling the absence of DISPLAYTITLE and we set talk page
title in BeforePageDisplay hook instead of ParserBeforeInternalParse and
ParserBeforePreprocess hooks.